Move create_main_worker to MainWorker::create (#6034)
