+++ cli/worker.rs
@@ -1,7 +1,10 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 use crate::fmt_errors::JSError;
+use crate::global_state::GlobalState;
 use crate::inspector::DenoInspector;
 use crate::ops;
+use crate::ops::io::get_stdio;
+use crate::startup_data;
 use crate::state::State;
 use deno_core::Buf;
 use deno_core::CoreIsolate;
@@ -246,7 +249,8 @@ impl DerefMut for Worker {
 pub struct MainWorker(Worker);
 
 impl MainWorker {
-  pub fn new(name: String, startup_data: StartupData, state: State) -> Self {
+  // TODO(ry) combine MainWorker::new and MainWorker::create.
+  fn new(name: String, startup_data: StartupData, state: State) -> Self {
     let state_ = state.clone();
     let mut worker = Worker::new(name, startup_data, state_);
     {
@@ -273,6 +277,35 @@ impl MainWorker {
       ops::worker_host::init(isolate, &state);
     }
     Self(worker)
+  }
+
+  pub fn create(
+    global_state: GlobalState,
+    main_module: ModuleSpecifier,
+  ) -> Result<MainWorker, ErrBox> {
+    let state = State::new(
+      global_state.clone(),
+      None,
+      main_module,
+      global_state.maybe_import_map.clone(),
+      false,
+    )?;
+    let mut worker = MainWorker::new(
+      "main".to_string(),
+      startup_data::deno_isolate_init(),
+      state,
+    );
+    {
+      let (stdin, stdout, stderr) = get_stdio();
+      let state_rc = CoreIsolate::state(&worker.isolate);
+      let state = state_rc.borrow();
+      let mut t = state.resource_table.borrow_mut();
+      t.add("stdin", Box::new(stdin));
+      t.add("stdout", Box::new(stdout));
+      t.add("stderr", Box::new(stderr));
+    }
+    worker.execute("bootstrap.mainRuntime()")?;
+    Ok(worker)
   }
 }
